backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
boto3==1.34.0
pandas==1.5.3  # Older compatible version
numpy==1.24.3  # Older compatible version
python-dotenv==1.0.0
aiofiles==23.2.1